split the classes in dt/synthesizer into two files; added decpomdp factory
