+++ paynt/dt/__init__.py
@@ -14,5 +14,6 @@ from .task import DtTask
 from .result import DtResult
 from .colored_mdp import DtColoredMdp
 from .factory import DtColoredMdpFactory
-from .synthesizer import SynthesizerARDt, DtSynthesizer
+from .synthesizer_ar_dt import SynthesizerARDt
+from .synthesizer import DtSynthesizer
 from . import dtnest
