+++ paynt/pomdp/decpomdp/__init__.py
@@ -1,2 +1,3 @@
-from .colored_mdp import DecPomdpColoredMdp, DecPomdpColoredMdpFactory
+from .colored_mdp import DecPomdpColoredMdp
+from .factory import DecPomdpColoredMdpFactory
 from .synthesizer import DecPomdpSynthesizer
